daemon status shows current wallet path
All the jsonrpc calls against electrum daemon is affected by the
implicit parameter of wallet in the cmd_runner.  AFAIK, there is no way
to check the value of this wallet.  This can be trouble some if multiple
wallets are loaded, or load_wallet command is executed several times.

This patch makes it viewable on the `daemon status`.